feat: TaskDetailView 구현 #3
